classify_service/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
google-cloud-storage
numpy
tensorflow-cpu==2.15.0
